# Optimize the performance of daxpy by using universal intrinsics