… getarch.c