## BackGround

"Universal Intrinsics" is  a set of APIs instead of a real unified set of intrinsics. This set of APIs is in a higher level than latter. Because it is oriented to the specific scenario of numerical computation, it can be easily converged to fewer common functions. Simultaneously, there is hardly a problem of no equivalent implementations on different platforms.
For developers, this mechanism can reduce the effort of adaptation to different platforms and still produce good performance. According to some test cases in OpenCV/Eigen/Numpy, the performance of the optimized functions is almost as good as the performance of manually optimized ones.  Here we optimized the Level-one daxpy function to demonstrate the power of USIMD.
## Benchmark

  Each test has run for 10 loops and then take the average result, The radio is the MFlops division of Baseline and optimized branch(*The bigger* the radio is, the *better* performance is achieved).

- X86 BaseLine(auto vectorized)

 | data size(10^3) | develop(float)               |
 | --------------- | ---------------------------- |
 | 4000            | 900.20 MFlops   0.008887 sec |
 | 8000            | 969.41 MFlops   0.016505 sec |

- X86-SSE2 enabled

  | data size(10^3) | usimd-daxpy(float)              | ratio |
  | --------------- | ----------------------------- | ----- |
  | 4000            | 1153.80 MFlops   0.006934 sec | 1.28  |
  | 8000            | 1606.26 MFlops   0.009961 sec | 1.65  |
  
- X86-FMA3 enabled

  | data size(10^3) | usimd-daxpy(float)              | ratio |
  | --------------- | ----------------------------- | ----- |
  | 4000            | 1187.21 MFlops   0.006738 sec | 1.31  |
  | 8000            | 1231.84 MFlops   0.012989 sec | 1.27  |
 
  
- X86-AVX2 enabled

  | data size(10^3) | usimd-daxpy(float)              | ratio |
  | --------------- | ----------------------------- | ----- |
  | 4000            | 2559.92 MFlops   0.003125 sec | 2.84  |
  | 8000            | 2559.96 MFlops   0.006250 sec | 2.64  |

## System Info

|           | x86                                      |
| --------- | ---------------------------------------- |
| Processor | Intel(R) Xeon(R) Gold 6161 CPU @ 2.20GHz |
| OS        | Windows Server 2008 R2 Enterprise        |
| Compiler  | MSVC14.06                                |

## Conclusion

- Compare with develop branch, when operating float dot, the new implementation has about 3x faster in AVX2, about 24%~65% faster in SSE2/SSE3. 
- Other architect such as NEON/VSX/MIPS/RISC-V  can be easily extended, just need to add a corresponding head file in `simd` module 

## Co-Author

@seiko2plus 

## Reference

1. https://software.intel.com/sites/landingpage/IntrinsicsGuide/
2. https://www.thetopsites.net/article/58234904.shtml
3. https://github.com/numpy/numpy/tree/master/numpy/core/src/common/simd
4. https://github.com/opencv/opencv/tree/master/modules/core/include/opencv2/core/hal