@martin-frbg 2020-09-23
Interesting concept. I wonder if this could be combined with the sse2neon.h from https://github.com/DLTcollab/sse2neon to port existing x86 optimizations to arm