@Qiyu8 2020-09-24
@martin-frbg of course, I'm working on another dot product function, which can easily extend to arm/power architect with the help of sse2neon, BTW, The opencv project is extending to [RISC-V](https://github.com/opencv/opencv/pull/18228) recently by using universal intrinsics.