@martin-frbg 2020-09-24
That is good to know - xianyi is already working on optimizing OpenBLAS for RISC-V (C-Sky 860 cpu, #2391) now that it has vector extensions.