@Diazonium 2020-09-24
I find it strange that you show results with FMA4 on an Intel system, no Intel CPU has ever supported FMA4.
Only the AMD Bulldozer series has FMA4, Zen1 has undocumented and possibly broken support, and Zen2 removed FMA4 completely.