@martin-frbg 2020-09-24
@diazonium Probably misnomer (SSE3/SSE4 ?) - the PR itself parallels their (accepted) contributions to numpy