@Qiyu8 2020-09-25
@Diazonium Thanks for pointing it out, I found that I actually running FMA3 benchmark two times,  as a response to your FMA4 question, please see [here](https://www.guru3d.com/news-story/fma4-instruction-set-hiddenbut-is-working-on-amd-zen-processors.html). Running an adapted script in the Zen processors that sends FMA4 instructions. The FMA4 task fired off at the processor surprisingly did not get refused and got executed successfully. And openblas has set the HAVE_FMA4 macro in Zen processors. so we will keep the HAVE_FMA4 instructions here.