@Diazonium 2020-09-25
> 
> 
> @Diazonium Thanks for pointing it out, I found that I actually running FMA3 benchmark two times, as a response to your FMA4 question, please see [here](https://www.guru3d.com/news-story/fma4-instruction-set-hiddenbut-is-working-on-amd-zen-processors.html). Running an adapted script in the Zen processors that sends FMA4 instructions. The FMA4 task fired off at the processor surprisingly did not get refused and got executed successfully. And openblas has set the HAVE_FMA4 macro in Zen processors. so we will keep the HAVE_FMA4 instructions here.

Yes Zen1 cores are able to execute FMA4 instructions, but we do not know why they are not advertised in the CPUID. It is possible that some of the instructions may trigger have rare corner cases where they are have bugs in Zen 1, which could manifest as precision issues or any other bugs.

I think it is a bad idea to enable any FMA4 codepath by default on Zen 1 cores, because we do not want to risk package maintainers shipping binaries that use undocumented instructions on some machines. Many people do not know or care about SIMD, because they are just using NumPy, SciPy, etc. that ends up using OpenBLAS, if FMA4 causes rare precision or stability issues.

I do think having FMA4 is great, but enabling it on Zen 1 cores should be disabled by default in Makefile.rule. Users willing to compile their own build of OpenBLAS would be free to experiment with turning it on.

And FMA 4 is definitely removed in Zen 2 cores.