@martin-frbg 2020-09-25
@Diazonium pretty sure HAVE_FMA4 is reported by getarch on Zen1 because the relevant CPUID bit is actually still set, but you are right in that e.g. Agner Fog considers it "not officially supported - do not use". Also pretty sure that none of the current ZEN-specific functions in OpenBLAS makes use of it.