@Qiyu8 2020-09-27
@martin-frbg @Diazonium Good advice, I have removed the default support for FMA4 instruments. we should leave this switch to users who want to try it out. 
@xianyi Glad to see RISC-V Vector supported on BLAS3, we can use universal intrinsics framework to support RISC-V on BLAS1, what do you think?